Compatible up to v3.0.5
I have recently reconfigured this file to be compatible with Sendy v3.0.5 and to make it easier to keep in sync. It will now support new features such as exclusions and segments, which were not available in the previous version. It will not include any bug fixes or new features introduced past v3.0.5 (until this file is updated again, obviously.)